ingestion/requirements.txt
fastapi==0.120.1
uvicorn[standard]==0.34.0
pydantic==2.10.5
pydantic-settings==2.7.1
httpx==0.28.1
qdrant-client==1.12.2
python-multipart==0.0.26
tiktoken==0.8.0
pypdf==6.10.2
python-docx==1.1.2
beautifulsoup4==4.12.3
lxml==6.1.0
markdown-it-py==3.0.0
python-magic==0.4.27
